Jorge Lara - Create logic, models

+++ src/main/java/com/portfolio/twitter/portfolio/repositories/PortfolioRepository.java
@@ -3,5 +3,5 @@ package com.portfolio.twitter.portfolio.repositories;
 import com.portfolio.twitter.portfolio.model.Portfolio;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PortfolioRepository extends CrudRepository<Portfolio, Long> {
+public interface PortfolioRepository extends CrudRepository<Portfolio, Integer> {
 }
