+++ src/main/java/com/portfolio/twitter/portfolio/service/PortfolioService.java
@@ -2,5 +2,5 @@ package com.portfolio.twitter.portfolio.service;
 
 import com.portfolio.twitter.portfolio.model.Portfolio;
 
-public interface PortfolioService extends CrudService<Portfolio, Long>{
+public interface PortfolioService extends CrudService<Portfolio, Integer>{
 }
